Show gotowe

+++ src/main/java/pl/coderslab/users/UserDelete.java
@@ -1,5 +1,6 @@
 package pl.coderslab.users;
 
+import pl.coderslab.utils.User;
 import pl.coderslab.utils.UserDao;
 
 import javax.servlet.*;
@@ -9,16 +10,27 @@ import java.io.IOException;
 
 @WebServlet( "/userDelete")
 public class UserDelete extends HttpServlet {
+
+    User[] userList;
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        int id=Integer.parseInt(request.getParameter("id"));
-        System.out.println(id);
 
-        UserDao userDao=new UserDao();
-        userDao.delete(id);
+        String verify = request.getParameter("ver");
 
-        response.sendRedirect("/userList");
-
+        if (verify!=null){
+            String sure=verify;
+            request.setAttribute("sure", sure);
+            UserDao userDao=new UserDao();
+            userList=userDao.findAll();
+            request.setAttribute("users", userList);
+            getServletContext().getRequestDispatcher("/users/list.jsp").forward(request,response);
+        } else {
+            int id = Integer.parseInt(request.getParameter("id"));
+            System.out.println(id);
+            UserDao userDao = new UserDao();
+            userDao.delete(id);
+            response.sendRedirect("/userList");
+        }
     }
 
     @Override
